run_model_on_device/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
av==13.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
flatbuffers==20181003210633
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.12.1
idna==3.10
Jinja2==3.1.4
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
ncnn==1.0.20240820
networkx==3.4.2
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
picamera2==0.3.21
pidng==4.0.9
piexif==1.1.3
pillow==11.0.0
portalocker==2.10.1
protobuf==3.20.3
psutil==6.1.0
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-prctl==1.8.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
scipy==1.14.1
seaborn==0.13.2
simplejpeg==1.7.6
six==1.16.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.23
ultralytics-thop==2.0.9
urllib3==2.2.3
v4l2-python3==0.3.5
Werkzeug==3.0.6
wrapt==1.16.0
